+++ robotica/types.py
@@ -1,0 +1,4 @@
+from typing import Any, Dict
+
+Action = Dict[str, Any]
+JsonType = Any